feat: update referral link, pass hex string to dataSuffix
